requirements.txt
python-dotenv
fastapi 
sqlalchemy 
uvicorn
pymysql
pydantic
pydantic-ai
